+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## Unreleased
+
+### Changed
+
+- 移除所有 Provider 中未使用的 `mergeCommonPlugins` 方法 (#1173)
+
+
 ## v3.8.0-beta.1 - 2026-05-12
 
 ### Fixed